python_bot/requirements.txt
telethon>=1.34.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
